requirements.txt
dash
pandas
pymysql
python-dotenv
openpyxl
plotly
dash-bootstrap-components
mariadb